Large-scale update. -Logger changed -Improved code readability -Added new logic of data processing -Improved error handling
The method of sending data to a third-party server has been updated and is now universal, it accepts the server address and the method to be sent with the data as parameters.

JWT tokens are now sent to the Redis server instead of the database.

Added function to get JWT tokens on request from the client side

More updates soon